Switch to new type system

+++ src/vm/result.rs
@@ -1,4 +1,3 @@
-use num_bigint::BigInt;
 use std::fmt;
 use std::fmt::Formatter;
 
@@ -54,15 +53,12 @@ impl RuntimeErr {
 
     pub fn new_item_does_not_exist<S: Into<String>>(
         type_name: S,
-        index: BigInt,
+        index: usize,
     ) -> Self {
         Self::new(RuntimeErrKind::ItemDoesNotExist(type_name.into(), index))
     }
 
-    pub fn new_item_cannot_be_set<S: Into<String>>(
-        type_name: S,
-        index: BigInt,
-    ) -> Self {
+    pub fn new_item_cannot_be_set<S: Into<String>>(type_name: S, index: usize) -> Self {
         Self::new(RuntimeErrKind::ItemCannotBeSet(type_name.into(), index))
     }
 
@@ -95,8 +91,8 @@ pub enum RuntimeErrKind {
     StringFormatErr(String),
     AttrDoesNotExist(String, String),
     AttrCannotBeSet(String, String),
-    ItemDoesNotExist(String, BigInt),
-    ItemCannotBeSet(String, BigInt),
+    ItemDoesNotExist(String, usize),
+    ItemCannotBeSet(String, usize),
     IndexOutOfBounds(usize),
     NotCallable(ObjectRef),
